Separate logic between AddonRepository and OrderedAddonViewModel
