+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/products/addons/AddonRepository.kt
@@ -1,7 +1,6 @@
 package com.woocommerce.android.ui.products.addons
 
 import com.woocommerce.android.annotations.OpenClassOnDebug
-import com.woocommerce.android.extensions.unwrap
 import com.woocommerce.android.model.Order
 import com.woocommerce.android.model.ProductAddon
 import com.woocommerce.android.model.toAppModel
@@ -16,29 +15,7 @@ class AddonRepository @Inject constructor(
     private val productStore: WCProductStore,
     private val selectedSite: SelectedSite
 ) {
-    companion object {
-        private const val regexGroupMatchSize = 3
-    }
-
-    private val orderAttributesKeyRegex = "(.*?) \\((.*?)\\)".toRegex()
-
-    private val String.asParsedPair
-        get() = orderAttributesKeyRegex
-            .findAll(this)
-            .first().groupValues
-            .takeIf { it.size == regexGroupMatchSize }
-            ?.toMutableList()
-            ?.apply { removeFirst() }
-            ?.let { Pair(it.first(), it.last()) }
-
-    fun mergeDataFromOrderedAddons(order: Order, productID: Long) =
-        fetchOrderedAddonsData(order, productID)
-            ?.unwrap { addons, attributes ->
-                attributes.mapNotNull { it.key.asParsedPair }
-                    .map { pair -> addons.find { it.name == pair.first } }
-            }
-
-    private fun fetchOrderedAddonsData(
+    fun fetchOrderedAddonsData(
         order: Order,
         productID: Long
     ): Pair<List<ProductAddon>, List<Order.Item.Attribute>>? =
